+++ stwo_cairo_prover/crates/prover/src/components/memory/memory_id_to_big/prover.rs
@@ -8,7 +8,9 @@ use rayon::iter::{
 use stwo_cairo_adapter::memory::{
     u128_to_4_limbs, EncodedMemoryValueId, Memory, MemoryValueId, LARGE_MEMORY_VALUE_ID_BASE,
 };
-use stwo_cairo_common::memory::{MEMORY_ADDRESS_BOUND, N_M31_IN_FELT252, N_M31_IN_SMALL_FELT252};
+use stwo_cairo_common::memory::{
+    split_f252_simd, MEMORY_ADDRESS_BOUND, N_M31_IN_FELT252, N_M31_IN_SMALL_FELT252,
+};
 use stwo_cairo_common::prover_types::simd::PackedFelt252;
 
 use super::component::{Claim, InteractionClaim, MEMORY_ID_SIZE};
@@ -16,7 +18,6 @@ use crate::cairo_air::relations;
 use crate::components::prelude::proving::*;
 use crate::components::range_check_vector::{range_check_9_9, SIMD_ENUMERATION_0};
 use crate::components::utils::{AtomicMultiplicityColumn, TreeBuilder};
-use crate::felt::split_f252_simd;
 
 pub type InputType = M31;
 pub type PackedInputType = PackedM31;
@@ -388,12 +389,11 @@ impl InteractionClaimGenerator {
 mod tests {
     use itertools::Itertools;
     use stwo_cairo_adapter::memory::{value_from_felt252, MemoryBuilder, MemoryConfig};
-    use stwo_cairo_common::memory::N_M31_IN_FELT252;
+    use stwo_cairo_common::memory::{split_f252, N_M31_IN_FELT252};
     use stwo_prover::core::backend::simd::m31::PackedM31;
     use stwo_prover::core::fields::m31::M31;
 
     use crate::components::memory::memory_address_to_id;
-    use crate::felt::split_f252;
 
     #[test]
     fn test_deduce_output_simd() {
